OSAC-3593: Add tests/__init__.py and fix stale imports for e2e relocation
PR #650 relocated e2e test suites under tests/e2e/ but left behind
stale import paths referencing the old tests/{suite}/ layout. This
commit:

- Creates tests/__init__.py (lost during git rename detection) so
  tests/ is a proper Python package again
- Updates all 22 files with stale imports to use tests.e2e.{suite}
  path
- Updates inline code docs to reference changed paths


rh-pre-commit.version: 2.4.0
rh-pre-commit.check-secrets: ENABLED
