+++ tests/e2e/caas/test_cluster_create.py
@@ -6,7 +6,7 @@ from pathlib import Path
 
 import pytest
 
-from tests.catalog.conftest import unique_name
+from tests.e2e.catalog.conftest import unique_name
 from tests.core.grpc_client import GRPCClient
 from tests.core.helpers import (
     wait_for_cluster_deleting,
