+++ tests/e2e/caas/test_cluster_delete_feedback_light.py
@@ -4,7 +4,7 @@ import contextlib
 import subprocess
 from pathlib import Path
 
-from tests.catalog.conftest import unique_name
+from tests.e2e.catalog.conftest import unique_name
 from tests.core.grpc_client import GRPCClient
 from tests.core.helpers import (
     wait_for_cluster_deleting,
