+++ tests/e2e/catalog/test_compute_instance_catalog_item_disk_image.py
@@ -4,7 +4,7 @@ import subprocess
 
 import pytest
 
-from tests.catalog.conftest import unique_name
+from tests.e2e.catalog.conftest import unique_name
 from tests.core.grpc_client import GRPCClient
 from tests.core.helpers import assert_grpc_rejected, wait_for_grpc_removal
