+++ tests/e2e/catalog/test_compute_instance_catalog_item_lifecycle.py
@@ -1,6 +1,6 @@
 from __future__ import annotations
 
-from tests.catalog.conftest import unique_name
+from tests.e2e.catalog.conftest import unique_name
 from tests.core.grpc_client import GRPCClient
 from tests.core.runner import poll_until
