+++ tests/e2e/vmaas/test_compute_instance_delete_during_provision.py
@@ -4,7 +4,7 @@ import time
 
 import pytest
 
-from tests.catalog.conftest import unique_name
+from tests.e2e.catalog.conftest import unique_name
 
 from tests.core.grpc_client import GRPCClient
 from tests.core.helpers import wait_for_cr, wait_for_deletion, wait_for_grpc_removal
