+++ tests/e2e/vmaas/test_compute_instance_gpu.py
@@ -4,7 +4,7 @@ import json
 import subprocess
 from typing import Any
 
-from tests.catalog.conftest import unique_name
+from tests.e2e.catalog.conftest import unique_name
 from tests.core.grpc_client import GRPCClient
 from tests.core.helpers import wait_for_cr, wait_for_deletion, wait_for_grpc_removal
 from tests.core.k8s_client import K8sClient
