+++ tests/e2e/vmaas/test_compute_instance_heartbeat.py
@@ -2,7 +2,7 @@ from __future__ import annotations
 
 import pytest
 
-from tests.catalog.conftest import unique_name
+from tests.e2e.catalog.conftest import unique_name
 
 from tests.core.grpc_client import GRPCClient
 from tests.core.helpers import wait_for_cr, wait_for_deletion, wait_for_grpc_removal, wait_for_provision, wait_for_running
