+++ tests/e2e/vmaas/test_compute_instance_restart_negative.py
@@ -3,7 +3,7 @@ from __future__ import annotations
 import time
 from datetime import UTC, datetime
 
-from tests.catalog.conftest import unique_name
+from tests.e2e.catalog.conftest import unique_name
 from tests.core.grpc_client import GRPCClient
 from tests.core.helpers import wait_for_cr, wait_for_deletion, wait_for_restart, wait_for_running
 from tests.core.k8s_client import K8sClient
