+++ tests/e2e/vmaas/test_compute_instance_stop_metering.py
@@ -5,7 +5,7 @@ from datetime import UTC, datetime
 
 import pytest
 
-from tests.catalog.conftest import unique_name
+from tests.e2e.catalog.conftest import unique_name
 from tests.core.grpc_client import GRPCClient
 from tests.core.helpers import (
     wait_for_cr,
