+++ tests/e2e/vmaas/test_console.py
@@ -16,7 +16,7 @@ from typing import Any
 import pytest
 import websocket
 
-from tests.catalog.conftest import unique_name
+from tests.e2e.catalog.conftest import unique_name
 from tests.core.grpc_client import GRPCClient
 from tests.core.helpers import (
     assert_grpc_rejected,
